Se realizo correcciones en el clase Vehiculo y Terrestre, tambien se crea las dos interfaces Motor y Vela
