+++ src/pk01/Vehiculo.java
@@ -1,25 +1,19 @@
 package pk01;
-public class Vehiculo {
-     protected int velocidadActual;
+abstract class Vehiculo {
+    protected int velocidadActual;
     protected int velocidadMaxima;
+
     public Vehiculo(int velocidadActual, int velocidadMaxima) {
         this.velocidadActual = velocidadActual;
         this.velocidadMaxima = velocidadMaxima;
     }
+
     public void imprimir() {
-        System.out.println("Velocidad Actual: " + velocidadActual + ", Velocidad Máxima: " + velocidadMaxima);
+        System.out.println("Velocidad Actual: " + velocidadActual + " km/h");
+        System.out.println("Velocidad Máxima: " + velocidadMaxima + " km/h");
     }
-    public void acelerar(int velocidad) {
-        velocidadActual += velocidad;
-        if (velocidadActual > velocidadMaxima) {
-            velocidadActual = velocidadMaxima;
-        }
-    }
-    public void frenar(int velocidad) {
-        velocidadActual -= velocidad;
-        if (velocidadActual < 0) {
-            velocidadActual = 0;
-        }
-    }
+
+    public abstract void acelerar(int velocidad);
+    public abstract void frenar(int velocidad);
 }
